example/requirements.txt
git+https://github.com/phanirithvij/deezloader@master
python-dotenv==0.10.3
